mental-health-support-agent/pyproject.toml
[project]
name = "mental-health-support-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "chainlit>=2.5.5",
    "openai-agents>=0.1.0",
    "python-dotenv>=1.1.1",
    "requests>=2.32.4",
]
